Change 'height' types from Long -> Int. Int.MaxValue is 2147483647, which is an order of magnitude more than current block heights on mainnet/testnet bitcoin (#685)

+++ core/src/main/scala/org/bitcoins/core/protocol/blockchain/ChainParams.scala
@@ -182,8 +182,8 @@ sealed abstract class ChainParams {
   /** In bitcoin [[MainNetChainParams mainnet]], the network recalculates the difficulty for the network every 2016 blocks
     * [[https://github.com/bitcoin/bitcoin/blob/eb7daf4d600eeb631427c018a984a77a34aca66e/src/consensus/params.h#L75 bitcoin core implementation]]
     * */
-  def difficultyChangeInterval: Long = {
-    powTargetTimeSpan.toSeconds / powTargetSpacing.toSeconds
+  def difficultyChangeInterval: Int = {
+    (powTargetTimeSpan.toSeconds / powTargetSpacing.toSeconds).toInt
   }
 
   /**
